added do enable thread restarts for data links

+++ src/main/java/ca/mrt/gs_backend/labjack/LabJackDataLink.java
@@ -199,14 +199,15 @@ public class LabJackDataLink extends AbstractTcTmParamLink implements Runnable{
     @Override
     protected void doStop() {
         if(isConnected){
+            executorService.shutdown();
             LJM.close(deviceHandle);
             isConnected = false;
+
             try {
                 csvWriter.close();
             } catch (IOException e) {
                 throw new RuntimeException(e);
             }
-
         }
         notifyStopped();
     }
@@ -222,12 +223,20 @@ public class LabJackDataLink extends AbstractTcTmParamLink implements Runnable{
             }
         }
 
+        initializeCSVWriterAndTasks();
+    }
+
+    private void initializeCSVWriterAndTasks(){
         try {
             File file = new File(CSV_FILENAME);
-            file.getParentFile().mkdirs();
-            log.info(file.getAbsolutePath());
-            csvWriter = new BufferedWriter(new FileWriter(file));
-            writeCSVHeader();
+            if(!file.exists()){
+                file.getParentFile().mkdirs();
+                log.info("Creating LabJack CSV file at: " + file.getAbsolutePath());
+                csvWriter = new BufferedWriter(new FileWriter(file));
+                writeCSVHeader();
+            } else{
+                csvWriter = new BufferedWriter(new FileWriter(file));
+            }
 
         } catch (IOException e) {
             throw new RuntimeException(e);
@@ -255,6 +264,8 @@ public class LabJackDataLink extends AbstractTcTmParamLink implements Runnable{
     @Override
     public void doDisable() {
         if (isConnected) {
+            executorService.shutdown();
+
             LJM.close(deviceHandle);
             isConnected = false;
             try {
@@ -264,6 +275,15 @@ public class LabJackDataLink extends AbstractTcTmParamLink implements Runnable{
             }
         }
     }
+
+    @Override
+    public void doEnable(){
+        Thread thread = new Thread(this);
+        thread.setName(getClass().getSimpleName() + "-" + linkName);
+        thread.start();
+    }
+
+
 
     @Override
     public String getDetailedStatus() {
